refactor(backend): decompose 837-line mod.rs into focused sub-modules
Extract 6 new modules from src/backend/mod.rs:
- panic_guard.rs: panic_safe() + PanicGuarded struct
- progress.rs: LspProgressReporter + KotlinProgress notification
- init.rs: workspace-root resolution + initialization methods
- capabilities.rs: server_capabilities() free function
- git_watcher.rs: read_git_commit + spawn_git_head_watcher
- commands.rs: execute_command_impl (reindex + clearCache)

mod.rs now contains only: module declarations, re-exports,
Backend struct + new(), and the thin LanguageServer dispatch impl.

Re-export panic_safe and LspProgressReporter from mod.rs to
preserve existing crate paths used in main.rs and tests.

All 1243 tests pass, zero clippy warnings.
